+++ AoC2022/Inputs/Input07.cs
@@ -35,9 +35,9 @@ public class Input07 : AbstractInput<IEnumerable<ICommandLineEntry>, IEnumerable
 
     protected override IEnumerable<ICommandLineEntry> ParseInputTwo(string input) => ParseInput(input);
 
-    protected override int SolveFirstPuzzle(IEnumerable<ICommandLineEntry> input) => Day07.FirstPuzzle(input);
+    protected override long SolveFirstPuzzle(IEnumerable<ICommandLineEntry> input) => Day07.FirstPuzzle(input);
 
-    protected override int SolveSecondPuzzle(IEnumerable<ICommandLineEntry> input) => Day07.SecondPuzzle(input);
+    protected override long SolveSecondPuzzle(IEnumerable<ICommandLineEntry> input) => Day07.SecondPuzzle(input);
 
     public override string TestInput =>
         "$ cd /\n$ ls\ndir a\n14848514 b.txt\n8504156 c.dat\ndir d\n$ cd a\n$ ls\ndir e\n29116 f\n2557 g\n62596 h.lst\n$ cd e\n$ ls\n584 i\n$ cd ..\n$ cd ..\n$ cd d\n$ ls\n4060174 j\n8033020 d.log\n5626152 d.ext\n7214296 k";
